// frontend/src/types/apps/carerTypes.ts
export type CarerType = {
    firstName?: string;
    lastName?: string;
    carerID?: any;
    emailID: string;
    password?: string;
    CarerID?: number;
    CarerName?: string;
    carerName?: string;
    profilePic?: File | string | null;
    descr: string;
    color: string;
    active: boolean;
    createdBy?: number | null;
    updatedBy?: number | null;
    createdOn?: string;
    updatedOn?: string;
    userID?: number | undefined;
    Active?: any;
    loading?: boolean;
    Descr: string;
    Email: string;
    CreatedByUserName: string;
    UpdatedByUserName: string;
    FirstName: string;
    LastName: string;
    EmailID: string;
    UpdatedOn: string;
    ProfilePic: any
    Color: string
    
}